backend + merging of frontend backend finished

+++ Frontend/src/AccountComponents/TransactionsDetails/TransactionsDetails.js
@@ -2,12 +2,15 @@ import classes from './TransactionsDetails.module.css'
 import NewFormModal from '../../Forms/ModalForm/NewFormModal';
 import { Fragment, useState, useEffect } from 'react';
 import NewForm from '../../Forms/AddNewForm/NewForm';
-
+import { getAuthToken } from '../../utils/auth';
 
 const TransactionsDetails = (props) => {
     const [ModalOpen, setModalOpen] = useState(false);
     const [formList, setFormList] = useState([]);
-    const [removeFormId, setRemoveFormId] = useState();
+    const [removeFormId, setRemoveFormId] = useState('');
+    const [isLoading, setIsLoading] = useState(true);
+    const token = localStorage.getItem('token');
+    const uid = localStorage.getItem('userId')
 
     const OpenModal = () => {
         setModalOpen(true);
@@ -15,29 +18,39 @@ const TransactionsDetails = (props) => {
     const CloseModal = () => {
         setModalOpen(false);
     };
-
     const deleteForm = (e) => {
         setRemoveFormId(e.currentTarget.id);
     };
-
-    useEffect(() => {
-        deleteDataHandler();
-        // getDataHandler();
-    },[removeFormId])
-
+    
 
     async function getDataHandler () {
-        const response = await fetch('https://money-management-5452c-default-rtdb.asia-southeast1.firebasedatabase.app/transactions.json')
+        const response = await fetch(process.env.REACT_APP_BACKEND_URL + `/form/${uid}/transactions`, {
+            method: 'GET',
+            headers: {
+                'Content-Type' : 'application/json',
+                'Authorization' : `Bearer ${token}`
+            }
+        });
         const myData = await response.json();
         const transactionsList = [];
         for (const key in myData){
+            if (myData[key].date){
             transactionsList.push({
-                id: key,
+                id: myData[key]._id,
+                date: myData[key].date.slice(0, 10).split('-').reverse().join('-'),
+                radioData: myData[key].radioData,
+                activityName: myData[key].activityName,
+                amountSpent: myData[key].amountSpent,
+                })
+            } else {
+                transactionsList.push({
+                id: myData[key]._id,
                 date: myData[key].date,
                 radioData: myData[key].radioData,
                 activityName: myData[key].activityName,
                 amountSpent: myData[key].amountSpent,
-            })
+                })
+            }
             
         };
         setFormList(transactionsList);
@@ -45,17 +58,31 @@ const TransactionsDetails = (props) => {
     };
 
     async function addDataHandler (data) {
-        const response = await fetch('https://money-management-5452c-default-rtdb.asia-southeast1.firebasedatabase.app/transactions.json', {
-            method: 'POST',
-            body: JSON.stringify(data),
-        });
+        try{
+            const response = await fetch(process.env.REACT_APP_BACKEND_URL + `/form/`, {
+                method: 'POST',
+                headers: {
+                    'Content-Type' : 'application/json',
+                    'Authorization' : `Bearer ${token}`
+                },
+                body: JSON.stringify(data),
+            });
+        } catch (e) {
+            console.log(e);
+        }
         getDataHandler();
     };
 
-
     async function deleteDataHandler (data) {
-        const response = await fetch(`https://money-management-5452c-default-rtdb.asia-southeast1.firebasedatabase.app/transactions/${removeFormId}.json`, {
+        if (removeFormId === ''){
+            return
+        }
+        const response = await fetch(process.env.REACT_APP_BACKEND_URL + `/form/${removeFormId}`, {
             method: 'DELETE',
+            headers: {
+                'Content-Type' : 'application/json',
+                'Authorization' : `Bearer ${token}`
+            },
             body: JSON.stringify(data),
         });
         getDataHandler()
@@ -65,7 +92,9 @@ const TransactionsDetails = (props) => {
         getDataHandler();
     },[])
 
-    console.log('refershed')
+    useEffect(() => {
+        deleteDataHandler();
+    },[removeFormId])
 
     const FormList = formList.map((form) => {
         return (
@@ -102,7 +131,7 @@ const TransactionsDetails = (props) => {
                     {FormList}   
                 </table>    
                     
-                {ModalOpen ? (<NewFormModal CloseModal={CloseModal} submitData={addDataHandler} getData={getDataHandler} radioDataNeeded={true}/>) : null}
+                {ModalOpen ? (<NewFormModal CloseModal={CloseModal} submitData={addDataHandler} getData={getDataHandler} radioDataNeeded={true} category={"transactions"}/>) : null}
             </div>
         </Fragment>
     );
